requirements.txt
# Core Framework
streamlit>=1.49.0
python-dotenv>=1.0.0

# AI & LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.3.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
langchain_huggingface>=0.3.0

# Search & Web Scraping
duckduckgo-search>=8.0.0
ddgs>=9.6.0
wikipedia>=1.4.0
arxiv>=2.2.0
beautifulsoup4>=4.13.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pypdf>=6.0.0
pymupdf>=1.26.0

# Vector Databases & Embeddings
faiss-cpu>=1.12.0
chromadb>=1.0.0
sentence-transformers>=5.1.0
huggingface_hub>=0.35.0

# Database & Storage
duckdb>=1.4.0
mysql-connector-python>=9.4.0
SQLAlchemy>=2.0.0

# Media & Content Processing
youtube_transcript_api>=1.2.0
unstructured>=0.18.0
pytube>=15.0.0

# Utilities
validators==0.28.1
numexpr>=2.12.0
ipykernel>=6.30.0

# Development & Testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0